+++ AtatusCore/Tests/Atatus/TracerTests.swift
@@ -77,7 +77,7 @@ class TracerTests: XCTestCase {
           "spans": [
             {
               "trace_id": "000000000000000a0000000000000064",
-              "span_id": "64",
+              "span_id": "0000000000000064",
               "parent_id": "0",
               "name": "operation",
               "service": "default-service-name",